+++ src/View/Helper/HtmlHelper.php
@@ -1,0 +1,75 @@
+<?php
+/**
+ * This file is part of cakephp-thumber.
+ *
+ * cakephp-thumber is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * cakephp-thumber is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with cakephp-thumber.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ */
+namespace Thumber\View\Helper;
+
+use Cake\View\Helper\HtmlHelper as CakeHtmlHelper;
+
+/**
+ * Html Helper.
+ *
+ * This class only rewrites the `HtmlHelper::image()` method.
+ */
+class HtmlHelper extends CakeHtmlHelper
+{
+    /**
+     * Creates a formatted `img` element for a thumbnail.
+     *
+     * Instead of `HtmlHelper::image()`, this method does not alter the image
+     *  path, making it possible to create the html tag for the thumb.
+     * @param string $path Thumbnail path
+     * @param array $options Array of HTML attributes
+     * @return string
+     */
+    public function image($path, array $options = [])
+    {
+        unset($options['fullBase']);
+
+        $options = array_diff_key($options, ['fullBase' => null, 'pathPrefix' => null]);
+
+        if (!isset($options['alt'])) {
+            $options['alt'] = '';
+        }
+
+        $url = false;
+        if (!empty($options['url'])) {
+            $url = $options['url'];
+            unset($options['url']);
+        }
+
+        $templater = $this->templater();
+        $image = $templater->format('image', [
+            'url' => $path,
+            'attrs' => $templater->formatAttributes($options),
+        ]);
+
+        if ($url) {
+            return $templater->format('link', [
+                'url' => $this->Url->build($url),
+                'attrs' => null,
+                'content' => $image
+            ]);
+        }
+
+        return $image;
+    }
+}